pyproject.toml
[project]
name = "mcp-demo"
version = "0.1.0"
description = ""
authors = [
    {name = "Codex",email = "codex@openai.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "openai (>=1.84.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "src"}]

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.12"
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
